Don't swallow exceptions when saving polymorphic models.
If an exception is raised in the `save()` method of a polymorphic model,
DDF tries to print all the field values for the model which will in turn
raise a `DoesNotExist` exception on the `*_ptr` field, that masks the
real exception.

This commit catches and displays the `repr()` of exceptions raised in
when displaying field values in the `print_field_values_of_a_model()`
method, which allows the real exception to exposed via `BadDataError`.

It might be more appropriate to just not display field values for
`*_ptr` fields, though, as they are not set directly by the user.