Reinstate tests for supported Django & Python versions, and fix tests for Python 3.2
Django 1.5 and 1.6 were not yet removed, even though they are not officially
supported anymore. Django 1.4 is also due to lose official support soon.

We should remove them from the envlist when it becomes to a burden to maintain
them or we want to cut down the number of environments being tested.

Removed the `except AssertionError` block in `test_ddf.py` at line 820 (but
kept the test). There's no point in running the tests if we are handling the
resulting AssertionError when the test fails.